Creating the first api endpoint.
The first api endpoint of the flask app will be: /api/frameworks and
it will server a list of objects and such objects are entities, in this
case, each entity represents a Framework (a piece of software).

We will dump our (python) list of Frameworks into a JSON that we will
serve at our application endpoint and we'll do it a in way in which
we'll fully understand the proccess of how it's done.