Creating the third api endpoint.
The second api endpoint of the flask app will be:
/api/frameworks and it will serve a (python) dictionary as
a JSON object.

This resource is served after creating a new entity based on
the JSON data that comes from a request made to this endpoint.

It's also worth to note that, in this endpoint, we're serving the
resource without explicitly calling jsonify. Also, it's important
to specify that, the only request methods accepted is: POST.