Creating the fourth api endpoint.
The fourth api endpoint of the flask app will be:
/api/frameworks/<int:id> and it will serve a (python)
dictionary as a JSON object.

This resource is served after trying to find a framework
that matches the id variable passed in the endpoint's URI,
if so then the framework's name is updated with the JSON data
passed in the PUT request; if not, then it will return a JSON
object with a message and a status_code with the request status
code of 404 ("Data not found").

In order to find a framework list comprehension is used and
some validations are made in order to retreive the correct
response data.