Creating the fifth api endpoint.
The fifth api endpoint of the flask app will be:
/api/frameworks/<int:id> and it will return no content as
a response.

No content is served after trying to find a framework
that matches the id variable passed in the endpoint's URI,
if so then the found framework is destroyed (deleted); if not,
then it will return a JSON object with a message and a
status_code with the request status code of 404 ("Data not found").

In order to find a framework list comprehension is used and
some validations are made in order to retreive the correct
response data. If a framework is successfully destroyed then no
content will be passed as response and the status code 204 is used.