Data persistency added by using SQLAlchemy.
Until this point, the REST API Flask WA has, at least, suffient
functionality to handle the common HTTP methods:

* GET
* POST
* PUT
* DELETE

Which are commonly known as methods for CRUD:

* Create
* Read
* Update
* Delete

But we're still using temporal data. In order to add data persistency
to the REST API we'll use SQLAlchemy, more specifically, we'll use
a package named: flask_sqlalchemy.

After installing flask_sqlalchemy, a SQLite3 database URI connection
is set and a SQLAlchemy model for the database is defined as
a class: Framework.

Now, we'll have to re-define our endpoints to implement the data
persistency through flask_sqlalchemy.