JWT auth implemented; requirements updated.
The app's code has received several changes in order to implement
authentication by using JWTAuth. Here are the most important changes:

* A new class User was created as a DB Model.
* A new class UserSchema was created as a schema for serializing and
loading user objects.
* The Framework and FrameworkSchema classes were modified since now
there's a relationship between frameworks and user objects which is
a foreign key called "author_id".
* A new POST API Endpoint was created ("/api/users") which will be used
  as an endpoint to sign up new users. Some validations are made in this
endpoint as well such as throwing errors when an user already exists.
* An application secret key was defined and it *MUST* be changed before
using it in a production environment.
* A new Endpoint was created ("/login") which is not necessarily
defined under the REST API Endpoints design. It'll be used for loggin-in
users and if there's success then a new JWT will be issued. Such JWT
will contain the user's id and username together with an expiration
timestamp that'll be used wherever the JWT needs validation. The JWT is
signed with the value of the application's secret key.

Now JWT Authentication is required for some endpoints:

* create_framework
* update_framework
* destroy_framewok

Because of that, now some new validations are made such as being able to
decode the JWT's contents successfully. Such validation will need code
refactorization under the DRY principle.

Changes for the file requirements.historical_refrence:

* Now there are references for the new package installed: pyjwt.