Code refactorization for JWT decodification.
The code refactorization that was mentioned at
72c0951a1f1da3381916c1e3efa5788438f0e2e4 has now been implemented in
order to apply the DRY principle.

A new function called jwt_required has been created and it'll serve as a
decorator for functions that need to decode a JWT Token.

The endpoints with the following functions have been updated:

* create_framework
* update_framework
* destroy_framework

Changes made to the mentioned functions:

* Those functions now are decorated with the function jwt_required and
they'll receive a decoded JWT and a user object. Those received
arguments will be used in the instructions of the functions to make
further validations.